+++ CHANGELOG.md
@@ -1911,3 +1911,4 @@
 - [2026-06-11 14:24] perf(state): apply RepaintBoundary on animated state widgets
 - [2026-06-11 14:40] fix(security): prevent double tap duplicate action in security
 - [2026-06-11 15:19] refactor(state): extract reusable atomic widgets from state
+- [2026-06-11 16:15] feat(i18n): add bi-directional event bus handler for i18n